feat: Add missing metadata
