+++ src/app/(layout-with-navigation)/(pages-statiques)/a-propos/page.tsx
@@ -1,22 +1,11 @@
-'use client'
+import { Metadata } from 'next'
+import APropos from './_components/APropos'
 
-import { getLocalisedMDX } from '@/helpers/getLocalisedMDX'
-import { useLocale } from '@/hooks/useLocale'
-import AboutEn from '@/locales/pages/en-us/about.mdx'
-import AboutFr from '@/locales/pages/fr/about.mdx'
-// import contentEs from '@/locales/pages/es/CGU.md'
-// import contentIt from '@/locales/pages/it/CGU.md'
+export const metadata: Metadata = {
+  title: 'À propos - Nos Gestes Climat',
+  description: 'Informations relatives à Nos Gestes Climat.',
+}
 
-export default function About() {
-  const locale = useLocale()
-
-  const AboutLocalised = getLocalisedMDX({
-    dictionnaries: {
-      fr: AboutFr,
-      'en-US': AboutEn,
-    },
-    locale: locale ?? '',
-  })
-
-  return <AboutLocalised />
+export default function AProposPage() {
+  return <APropos />
 }
