+++ src/app/(layout-with-navigation)/(pages-statiques)/budget/page.tsx
@@ -1,44 +1,11 @@
-'use client'
+import { Metadata } from 'next'
+import Budget from './_components/Budget'
 
-import { getLocalisedMDX } from '@/helpers/getLocalisedMDX'
-import { useLocale } from '@/hooks/useLocale'
-import contentEnBottom from '@/locales/pages/en-us/budgetBottom.mdx'
-import contentEnTop from '@/locales/pages/en-us/budgetTop.mdx'
-import contentFrBottom from '@/locales/pages/fr/budgetBottom.mdx'
-import contentFrTop from '@/locales/pages/fr/budgetTop.mdx'
-import SelectYear from './_components/SelectYear'
+export const metadata: Metadata = {
+  title: 'Budget - Nos Gestes Climat',
+  description: 'Informations relatives au budget de Nos Gestes Climat.',
+}
 
-export default function Budget() {
-  const locale = useLocale()
-
-  const BudgetContentTop = getLocalisedMDX({
-    dictionnaries: {
-      fr: contentFrTop,
-      'en-US': contentEnTop,
-    },
-    locale: locale ?? '',
-  })
-
-  const BudgetContentBottom = getLocalisedMDX({
-    dictionnaries: {
-      fr: contentFrBottom,
-      'en-US': contentEnBottom,
-    },
-    locale: locale ?? '',
-  })
-
-  return (
-    <>
-      {/*
-      <Meta
-				title={t('meta.publicodes.Budget.title')}
-				description={t('meta.publicodes.Budget.description')}
-			/>
-      */}
-
-      <BudgetContentTop />
-      <SelectYear />
-      <BudgetContentBottom />
-    </>
-  )
+export default function BudgetPage() {
+  return <Budget />
 }
