+++ src/app/(layout-with-navigation)/(pages-statiques)/diffuser/page.tsx
@@ -1,22 +1,11 @@
-'use client'
+import { Metadata } from 'next'
+import Diffuser from './_components/Diffuser'
 
-import DiffuserEn from '@/locales/pages/en-us/diffuser.mdx'
-import DiffuserFr from '@/locales/pages/fr/diffuser.mdx'
+export const metadata: Metadata = {
+  title: "Diffuser notre simulateur d'empreinte climat - Nos Gestes Climat",
+  description: 'Diffusez Nos Gestes Climat dans votre organisation.',
+}
 
-import { getLocalisedMDX } from '@/helpers/getLocalisedMDX'
-
-import { useLocale } from '@/hooks/useLocale'
-
-export default function Diffuser() {
-  const locale = useLocale()
-
-  const DiffuserLocalised = getLocalisedMDX({
-    dictionnaries: {
-      fr: DiffuserFr,
-      'en-US': DiffuserEn,
-    },
-    locale: locale ?? '',
-  })
-
-  return <DiffuserLocalised />
+export default function DiffuserPage() {
+  return <Diffuser />
 }
