+++ src/app/(layout-with-navigation)/(pages-statiques)/guide/[category]/page.tsx
@@ -1,38 +1,20 @@
-'use client'
-
 import Trans from '@/components/translation/Trans'
 import ButtonLink from '@/design-system/inputs/ButtonLink'
-import { handleFetchDocumentation } from '@/helpers/fetchDocumentation'
-import { getRuleTitle } from '@/helpers/publicodes/getRuleTitle'
 
-import { NGCRule } from '@/publicodes-state/types'
-import { useParams } from 'next/navigation'
-import { utils } from 'publicodes'
-import { use } from 'react'
-import { getCategoryContent } from '../_helpers/getCategoryContent'
+import { Metadata } from 'next'
+import CategoryGuide from './_components/CategoryGuide'
 
-export default function GuideAlimentation() {
-  const documentationData: { [key: string]: string } = use(
-    handleFetchDocumentation
-  )
+export const metadata: Metadata = {
+  title: 'Le guide - Nos Gestes Climat',
+  description:
+    'Retrouvez dans ce guide toutes les informations sur Nos Gestes Climat.',
+}
 
-  const { category } = useParams()
-
-  const Content = getCategoryContent(category as string)
-
-  if (!Content) return null
-
-  const actionsPlus = Object.entries(documentationData)
-    .filter(([key]) => key.startsWith('actions-plus/'))
-    .map(([key, value]) => ({
-      plus: value,
-      dottedName: key.replace('actions-plus/', ''),
-    }))
-
-  const relatedActions = actionsPlus.filter(
-    (action) => category === action.dottedName.split(' . ')[0]
-  )
-
+export default function CategoryGuidePage({
+  params: { category },
+}: {
+  params: { category: string }
+}) {
   return (
     <div className="mx-auto my-4 flex flex-col items-start justify-center">
       <ButtonLink color="text" href="/guide">
@@ -40,29 +22,7 @@ export default function GuideAlimentation() {
         <Trans>Retour</Trans>
       </ButtonLink>
 
-      <div>
-        <Content />
-
-        <h2>
-          <Trans>Pour aller plus loin</Trans>:
-        </h2>
-        <div>
-          {relatedActions.map((action) => (
-            <ButtonLink
-              key={action.dottedName}
-              color="secondary"
-              size="sm"
-              href={'/actions/plus/' + utils.encodeRuleName(action.dottedName)}>
-              {getRuleTitle(
-                action as unknown as NGCRule & {
-                  dottedName: string
-                  titre: string
-                }
-              )}
-            </ButtonLink>
-          ))}
-        </div>
-      </div>
+      <CategoryGuide category={category} />
     </div>
   )
 }
