+++ src/app/(layout-with-navigation)/(pages-statiques)/guide/_components/Guide.tsx
@@ -1,0 +1,7 @@
+'use client'
+
+import GuideContent from './_content/guide.mdx'
+
+export default function Guide() {
+  return <GuideContent />
+}
