+++ src/app/(layout-with-navigation)/(pages-statiques)/nouveautes/page.tsx
@@ -1,7 +1,14 @@
 import Title from '@/design-system/layout/Title'
 
 import { getServerTranslation } from '@/helpers/getServerTranslation'
+import { Metadata } from 'next'
 import NewsList from './_components/NewsList'
+
+export const metadata: Metadata = {
+  title: 'Les nouveautés - Nos Gestes Climat',
+  description:
+    'Consultez les nouvelles fonctionnalités et dernières nouvelles de Nos Gestes Climat.',
+}
 
 export default async function News() {
   const { t } = await getServerTranslation()
@@ -11,7 +18,9 @@ export default async function News() {
   return (
     <>
       <Title data-cypress-id="news-title" title={title} />
+
       <p>{t('pages.News.premierParagraphe')}</p>
+
       <NewsList />
     </>
   )
